Refactors the contact page
